Effects Cutter, by Sreejesh
Adding Effects Cutter script, by Sreejesh:
"This is a small script based on the Auto Talkback script. This works as
the reverse of it. This Plug-in will act like a Effects Bleed control
based on the Transport in Pro Tools. If it is in play, the bleed will be
open and when it stops, the bleed will close. This is a useful tool if
you don’t want to hear the tail of the reverb or delay whenever you stop
transport. This plug-in will cut it based on the bleed percentage. If it
is set to 0, then the reverb tail will be completely cut. Else, it will
allow whatever level is set in it. If set to 100 %, the plug-in will be
effectively like a bypass. In a signal chain, this will come after
whatever effects you have on the aux or track."